create_svgz {svglite}	R Documentation

Convert an svg file to svgz, overwriting the old file

Description

Convert an svg file to svgz, overwriting the old file

Usage

create_svgz(file)

Arguments

file
the path to the file to convert

[Package svglite version 2.1.2 Index]